+++ drizzle/0002_mute_jackal.sql
@@ -1,0 +1,13 @@
+CREATE TYPE "public"."invite_status" AS ENUM('pending', 'accepted', 'rejected');--> statement-breakpoint
+CREATE TABLE "invites" (
+	"id" text PRIMARY KEY NOT NULL,
+	"creator_id" text NOT NULL,
+	"email" text NOT NULL,
+	"status" "invite_status" DEFAULT 'pending' NOT NULL,
+	"token" text NOT NULL,
+	"created_at" timestamp DEFAULT now() NOT NULL,
+	"updated_at" timestamp DEFAULT now() NOT NULL,
+	CONSTRAINT "invites_token_unique" UNIQUE("token")
+);
+--> statement-breakpoint
+ALTER TABLE "invites" ADD CONSTRAINT "invites_creator_id_user_id_fk" FOREIGN KEY ("creator_id") REFERENCES "public"."user"("id") ON DELETE cascade ON UPDATE no action;